Fix crash in ValueToInterpValue when the Value is a tuple and the optional ConcreteType is a struct.
ValueToInterpValue() takes an optional ConcreteType that can attach sign information to an InterpValue that is not present in the Value. There is some logic that matches ConcreteType to ValueKinds to see through arrays and tuples, but structs were not properly handled. This change handles structs correctly and adds test coverage.

PiperOrigin-RevId: 551572500