[CI] Add a build_test to check DSLX .x test files have an associated target.
This ensures we don't miss things running in CI by accident.

In the process we also massage some of the `.x` files to have a main entry
point, so that `convert_to_ir` can be true and trigger the IR conversion path
in the `dslx_lang_test` macro and associated mechanisms like Z3 conversion.

Special thanks to tedhong@, allight@, and epastor@ for providing a set of potential
solutions that were nicer than a github action, in terms of "not living outside
of the normal testing flow".

An alternative proposal was globbing .x files and .ir files and passing them to
a script for cross checking.

PiperOrigin-RevId: 551589303