Make code template easier to use: provide replacements via callback.
The template filling was a bit hard to use, requiring to fill an
array of a particular size with all the substitutions.

Instead, make this now more on-demand: the Substiute() function
will ask the provided callback for replacments. This feels more
'natural to code and makes it both easier to use and more flexible.

While at it: we're not using CodeTemplate in std::optional anymore,
so the assignment operator can go, encouraging the user not to
copy this object.

PiperOrigin-RevId: 551693803