Allow to use DSLX template parameters in FFI templates.
Template parameters often contain crucial information that needs
to be passed to foreign function invocations. A typical example
would be parameters like BIT_WIDTH that need to be passed to
Verilog #parameters.

Constant expressions are already eliminated at the time the
IR arrives at the code generation, so here we replace constant
values early in the process and partially fill the template with
these values before passing them down.

PiperOrigin-RevId: 551731092