Add analysis of which predicates instructions are dominated by.
This is the PredicateDominatorAnalysis analysis for determining the scope of selects zones of control. The analysis will report, for every node, which select-branch guards use of that node.

Currently only the final select is actually recorded and reported since in most cases the combinatorial explosion of the nested cases make the full case list irrelevant. This could easily be added later however if it proves useful.

Bug: #1059
PiperOrigin-RevId: 551865997