Reserve space in a hash-set for performance of TopoSort
This improves speed substantially. Depending on the size and structure of the input anywhere from a 5% (20-deep balanced binary tree) to 50% (512-deep 32-wide dense network) improvement in topological sorting.

Note that these benchmark ir graphs are somewhat atypical. The most normal is the 'ladder' ir which has ~25% improvement but even that is more connected than typical graphs. For more normal uses the result will likely be less dramatic

PiperOrigin-RevId: 552509907