xls:simulation:generic: Add proto definition for simulation config xls:simulation:generic: Add configuration wrappers xls:simulation:generic: Add config tests
This commits adds proto3 definition for configuration
file used for simulation.

It also adds wrapper that abstracts from DLSX/IR simulation.
It takes the simulation type as a constructor argument and
manages channels based on target simulation type.

It adds tests for the configuration wrappers
for the renode simulation proto config.

Internal Tag: [#47603]

Co-Authored-by: Pawel Czarnecki <pczarnecki@antmicro.com>
Co-Authored-by: Krzysztof Boronski <kboronski@antmicro.com>

Signed-off-by: Maciej Dudek <mdudek@antmicro.com>
Signed-off-by: Krzysztof Boronski <kboronski@antmicro.com>