xls:simulation:generic: Add IChannel
Add IChannel interface definition.
Ichannel returns status access methods, to provide the ability
to handle errors on the lowest possible level without the need of checking
if the method usage was valid by external classes.

It also substitutes a pure virtual destructor with a virtual destructor,
because in the case of this class, it simplifies the codebase without
changing the operation of the class.

Internal-tag: [#47333] [#47334] [#47581]

Co-authored-by: Robert Winkler <rwinkler@antmicro.com>

Signed-off-by: Roman Dobrodii <rdobrodii@antmicro.com>
Signed-off-by: Robert Winkler <rwinkler@antmicro.com>