Add predicate-context api to query-engine
These will be used to implement context-sensitive range analysis. Currently the context is ignored by all engines. This behavior is fully consistent with the API since the set of facts that are true about X must be a superset of the set of facts which are true about X given Y.

Bug: google/xls#1059
PiperOrigin-RevId: 553569693