Make sure we schedule `send`s strictly before any `recv`s that depend on them
This makes our semantics compatible with our current codegen, which currently blocks all side effects (including `send`) of a stage on any blocking `recv`s that might be scheduled in the same stage.

Before this, code of the form:
```
tokA = send(tok, ...)
... = recv(tokA, ...)
```
could end up with the `send` and `recv` scheduled in the same stage, meaning that if the `recv` blocked, it would block the `send`. This could result in deadlocks in scenarios where the `recv`'s channel will only be populated in response to the `send`.

We might consider alternatives involving changes to codegen in the future.

PiperOrigin-RevId: 553944088