Add pass that rewrites assertion conditions.
Previously, assertions would fire even if the inputs to the assertion were not valid. This could lead to all sorts of trivial assertion failures on reset as well as other spurious failures.

This change to codegen emits assertions as `!pipeline_valid[stage] || condition` for functions and `!stage_done[stage] || condition` for procs, ensuring the assertions are only checking `condition` when it is valid.

PiperOrigin-RevId: 554658202