Determine the minimum pipeline length with SDC scheduling
ASAP scheduling (as expressed by ScheduleBounds) does not account for I/O and other scheduling constraints. Before this change, this could cause us to fail if the block would fit in a shorter pipeline without the constraints. For examples: see the new tests in `pipeline_schedule_test.cc`, which exercise the default `SendThenRecvConstraint` more thoroughly.

Therefore, if the pipeline length is not specified with SDC scheduling, we use the full scheduling model to find the minimum pipeline length. As part of this change, we need to ensure that the SDC scheduler applies all of our original constraints while constructing its optimization problem, including that function params must be in stage 0, function return values must be in the last stage, and no single node can have a delay greater than our cycle length.

PiperOrigin-RevId: 555225541