[DSLX:parser] Refactor to consolidate on AST cloner.
Types needed slightly shallower cloning to share the cloner infrastructure, if
we recursed all the way into the type definitions (beyond the references) then
we'd break the nominal type comparisons we do by AST pointer. This adds a
cloning option that doesn't traverse into type definitions, which is what we
need for return type clones as the parser currently does them.

PiperOrigin-RevId: 555515229