Optimize common next-state element access patterns
We track & retain the next-state mapping from Node* to the relevant state element indices, reducing the times we need to iterate over all state elements just to determine that our Node* is not a next-state element for any of our state elements.

While we're at it, we also replace a loop over all (node, state element) pairs with a simpler loop over the state elements.

This ends up saving ~100s in an internal slow codegen example, or about 30% of runtime (after a previous change cut runtime by ~70%).

PiperOrigin-RevId: 555699117