Update Python API `generate_pipelined_module_with_n_stages`: If a pipeline schedule with 1 stage is requested, just do the simple thing and put all the nodes into a single stage.  This removes any dependency on the delay model for this case and fixes https://github.com/google/xls/issues/1015.
When using this simplification with the Python API, some of the new detailed
scheduling error checks are skipped, for example, for rare cases that would make
a single-stage pipeline infeasible such as send/receive with minDelay > 0.
We don't expect this API to handle such complex cases since it doesn't expose
the scheduling or codegen options.

PiperOrigin-RevId: 556101627