[DSLX:ast] Make proc members parse in to a ProcMember AST node.
Previously they were "params" and they were kind of aliased into parameters
held on the configuration function, which felt like a strange and confusing
representational structure (particularly when things went wrong in another
change I was trying to make!). Now they are explicitly represented in the AST
as proc members.

Some of the "funkiness" comes from the fact that in the `config` function for
procs the *parameters* are *constexpr* (which is not typical of parameters,
usually they are runtime supplied!) -- this is because config/spawn proc
network instantiation all happens at constexpr configuration time.

PiperOrigin-RevId: 556108546