[DSLX] Rip out proc member transaction that parsed proc member tokens 2x.
This meant we needed to properly clone the types for the params in proc config
functions so they would have a correct parent/child relationship, instead of
sneakily borrowing them from the ProcMembers.

Also realized we didn't check if proc functions are duplicated, so added an
easy little inline check for that.

Also cleaned up the logic for detecting references to proc members in the
config function so it reads more linearly via helper functions.

PiperOrigin-RevId: 556836241