[ir:value] Hoist BitsToString to bits_ops.h
This is a pre-factoring that will allow us to support bigint conversions to
string, which are currently unimplemented.

By moving the stringification methods out to on-the-wall-functions in
bits_ops.h we can use BigInt and keep the layering where bits.h is its own
translation unit wrapping an InlineBitmap. Otherwise this would all have to
collapse into a mega-translation unit. We don't use the `.ToString()` method in
/that/ many places, and this introduces AbslStringify for the fairly common
case of StrFormat/StreamFormat usage.

PiperOrigin-RevId: 557302497