Keep the fuzzer from trying to run samples through codegen with too-small pipeline depth
This should fix some issues we've seen since adding the default constraint that a recv depending on a send must be scheduled in a strictly-later stage (due to current limitations of codegen).

Unfortunately, this change is extremely invasive, since we need to track the possible impact of depending on a send through all expressions we handle.

PiperOrigin-RevId: 557632968