[DSLX] Warn on unused variables -- examples/modules/xls_go_math
Prep for enabling unused warnings-as-errors.

This found an typo in the match statements in the riscv_simple example, which
bound the matched expression to the typo! This is exactly the kind of mistake
this warning is meant to flag.

Note: a leading underscore indicates "this is unused, but that's ok, I still
want to give some identifier information about it".
PiperOrigin-RevId: 558282375