Create reusable SDC scheduler to save computation
This lets us save time during the binary search we use to minimize the clock period (if unspecified), and lets us reuse that solver to speed up the final optimization as well.

This reduces the runtime spent in codegen by ~20% on one slow internal example, and in scheduling specifically by ~50% on our internal benchmark suite.

PiperOrigin-RevId: 558796321