[DSLX] Plumb a warning-enabled-set everywhere.
This is prep for adding a warn-on-unused with warnings-as-errors. There may be
some targets we want to temporarily opt out of the warning while landing it,
then flip the flag for all the targets.

Plumbs warnings around as a flag set and adds an interpreter test that we can
disable a particular warning in `interpreter_main`. The warning collector
discards any flagged warnings that are in the disabled set. The warning names
are a bit long but that's ok, people should rarely-if-ever disable them.

Also since WarningCollector is no longer default-constructable
TypecheckedModule is transitively not default-constructible so get it returned
as a statusor vs populating it as an outparam in some tests.

PiperOrigin-RevId: 558800148