[DSLX:parser] Create a positional error on collision at module scope.
Previously this would create a non-positional error.

Note that we keep the AST separately layered from the ParseError definition by
layering in an error-making strategy via a lambda. In generator code, like the
fuzzer's AST generator or the proto2dslx code, it knows because it's constructing
the module that an error should not be raised, and it can deal with a
non-positional InvalidArgumentError.

Found via coverage-guided fuzzing.

PiperOrigin-RevId: 559546483