[IR:opt] Label (for cover & assert) recovery pass.
Previously for coverpoints we would get very terribly mangled names. This was
because they /could/ end up conflicting. However, most of the time, they do not
conflict, so this introduces a relatively simple pass that recovers the
original labels if there were no collisions in the inlining process (applying
the pass at the end of the standard pipeline).

PiperOrigin-RevId: 560170135