Improve handling of Gate op by narrowing and constant-folding
These passes treated Gate as side effecting and avoided doing anything to it (mostly, narrowing would remove gates in some circumstances). Changed both of these passes to treat gate as a standard instruction. This is desirable because gate is meant to reduce power consumption and replacing it with a constant works far better than any sort of switching.

Updated docs a bit to make Gate's purpose clearer.

PiperOrigin-RevId: 565704407