[solvers] Add support for conjunctions of predicates in Z3 IR translation.
This is a useful facility when we're trying to make proofs that ranges are
precisely correct; i.e. over multiple sub-ranges of the form AND(2<=x, x<=4)
and multiple return values. This is infrastructure pre-work for a
range-analysis formal fuzzer.

Implementation note: we support conjunctions of predicates by extending the way
we handle single predicates. For single predicates P, we ask Z3 to find a
counterexample to the assertion ~P, and if it comes up unsat (no counterexample
can be found), predicate P is proven in proof-by-contradiction style. Similarly
now, for multiple predicates, P, Q, R our composite assertion is C = P & Q & R, so
we apply modus ponens on C and enqueue ~P | ~Q | ~R to the solver to attempt to
find a contradiction where any of the terms has a counterexample.

This change has small niceities used in bringing up this functionality:

* Source locations on the Z3 ScopedErrorHandler to more precisely pinpoint
  where things went wrong in the caller source text (vs just getting a logging
  line number in the utility library).
* Unsigned-less-than and unsigned-greater-than op translations that supplement
  the existing comparison ops.
* Some boolean helpers -- Z3 assertions work with boolean values (not
  bit-vectored values) so these were more necessary for developing the
  conjunction support.
* Renaming the predicate's EqualTo(node) to IsEqualTo(node) for a bit of naming
  consistency -- subject/object separation is a bit more clear now with the
  PredicateOfNode structure.

PiperOrigin-RevId: 568667124