[DSLX] Parenthesize LHS when cast is used as LT operand.
Without parenthesization thiscreates a situation where the formatted result
cannot be parsed, as we think that in something like:

    x as t < x
         ^~~~^

The open angle is potentially starting a parameterized type instantiation of
`t` in the grammar.

This was observed when making fuzzer modifications that generate this construct
(by constructing raw AST nodes and then formatting them).

PiperOrigin-RevId: 569002585