[DSLX:fmt] Use pretty printing library on first AST nodes.
Establishes plumbing we have to fill in to get nodes enqueued into the pretty
printing reflow algorithm and gives a case study with a `let` AST node that has
an inline comment (that the Let node is nominally responsible for emitting).

The Let node create a pretty printable doc that represents its syntax, then
sees if it is responsible for a comment -- if so, attempts to emit it inline.
If it doesn't fit inline, puts it in the line above the statement.

Note that this does not yet handle the case where there is a comment
immediately above the Let being associated with the Let, but it gets us off the
ground and establishes the basic scaffolding.

(Aside: before this tool is used in "mission mode" all fatals must be
eliminated -- they just serve as a convenient mechanism to get the plumbing all
set up and start filling in details incrementally in CLs.)

PiperOrigin-RevId: 570160167