Add NodeDependencyAnalysis
This analysis will determine which nodes can affect one another (either forward, answering the question 'could a change in X cause a change in Y', and backwards, answering the question 'could a change in X have been caused by a change in Y'). These can be used to define event horizons where values are known to be independent of one another. This is useful in some optimizations to avoid unneeded work.

Efforts are made to optimize this analysis to ensure it completes relatively quickly, including by allowing one to request the analysis only collect information on a subset of nodes.

This analysis both returns a struct which can answer 'IsDependent(A, B)' questions as well as return the full bit-map of dependencies keyed on a nodes id.

PiperOrigin-RevId: 570712266