Add instructions on how to add additional caching to XLS builds
We've realized that different developers were using different solutions for this, including `ccache` - so we tried out a few options, and found that Bazel's built-in cache support seems to be the most effective (~10x faster than `ccache` for a clean build with hot cache). `--disk_cache` works best with minimal setup, but localhost instances of `bazel-remote` reduce the need for maintenance.

PiperOrigin-RevId: 571995943