[DSLX:fmt] Add tests that exercise paren-insertion path.
This can generally only be exercised by constructing custom AST data structure
and formatting it (as the AST generator library does in the fuzzer). Reason
being: to get a similar construct to parse the same way from text, you'd need
to wrap it in parentheses, and the AST retains input parentheses on formatting.

Here we factor out a snippet that was used for the previous (non
pretty-printing) formatter for use in both test cases.

PiperOrigin-RevId: 572038135