Add --llvm_jit_ir_output and --llvm_jit_asm_output args to eval_ir_main
One feature of eval_ir_main is its ability to print out the LLVM ir and asm for compiled modules. This adds support to eval_ir_main for sending these directly to various files.

This is done by adding a basic observer system to the jit which is notified when compilation is performed.

PiperOrigin-RevId: 572435953