[DSLX:fmt] Allow comments within function scope.
Also fix that hard lines between indented constructs would leave trailing
whitespace -- to do this we introduce a notion of "real output column" (i.e.
where we emitted text) and "virtual output column" (i.e. where we'd be emitting
text to if we had inserted leading spaces eagerly).

This is also kind of nice as it makes emit() and emit_cr() helper routines
inside the pretty printer machinery to encapsulate the pushing of string pieces
into the resulting doc.

PiperOrigin-RevId: 574518984