common/subprocess: predecl environ
per man(7) environ:
```
Historically and by standard, environ must be declared in the
user program.  However, as a (nonstandard) programmer
convenience, environ is declared in the header file <unistd.h> if
the _GNU_SOURCE feature test macro is defined.
```

It turns out to be required on macos platforms where unistd.h doesn't declared it.

PiperOrigin-RevId: 574856323