Revert automatic "fix" that broke the build
An internal code-fixing bot suggested that, since the AddLogSink and RemoveLogSink functions were not included in any header used by their defining source file, they should be marked static so they would have internal linkage. The actual issue is that the header that declares these functions is not included in the file that defines them. It's tricky to actually include it without creating a dependency loop, so we'll figure that out in a followup change.

PiperOrigin-RevId: 575828050