Handle all global LogSink interactions in a cc_library
We were previously defining AddToSink & RemoveToSink in a source file that could not see their declarations in logging.h - and fixing this would have produced a cyclic dependency. Instead, we split off a log_sink_registry library that can handle all the interactions with global LogSinks, avoiding the loop.

This finishes the cleanup after commit 3033875c.

PiperOrigin-RevId: 575938205