[DSLX:fmt] More uniform handling of inline comments on let statements.
Before we were trying to split multiple-inline-comment from one-inline-comment
cases. This starts with proper handling of real world use cases of single
inline comments, and we'll correct from here.

The single inline comment is reflowed up the text width of the pane.

This also corrects for some erroneous spaces that were being placed at the
start of blocks.

PiperOrigin-RevId: 576664789