Codegen: Provide an actionable error message if function call found in IR.
The codegen assumes that the IR has been through the optimizer first, i.e. does
not contain any function calls anymore -- except FFI invocations.

The error message in that case was pointing out this technical aspect.
Given that this should never happen in the usual case, but is likely to happen
if codegen is called with un-optimized IR, adapt the error message
to inform how we likely got here and how to fix.

Issues: #1170
PiperOrigin-RevId: 577001122