[DSLX:frontend] Do not have the parser resolve struct references.
The fact we did this made it hard to format, as the AST didn't have fidelity to
the original source, and it didn't really make sense to be doing at this point
as type inference will make these associations.

As fallout, modifications to the directly-AST-constructing tools like
proto2dslx and the fuzzer were needed.

PiperOrigin-RevId: 577272195