[DSLX:fmt] Fix parentheses determination in autofmt.
Previously we would use the public `GetPrecedence()` value when formatting an
op, which says "strongest" when the operation is wrapped in parentheses by the
user. For formatting of an operation we want to know if we should parenthesize
*arguments* vs the natural precedence of the *operator*, so we want to query
the operation's natural precedence (without parens).

Renames GetPrecedenceInternal() to GetPrecedenceWithoutParen() and makes it
public for this use case.

PiperOrigin-RevId: 577292025