[DSLX:fmt] End-of-line comments were preventing paragraph-style spacing.
This updates inline comments to reflect that they end on the line the start on.
The basic source of difficulty is that a comment's text includes the newline
character, and thus its limit position is the 0 column on the subsequent line.
However, for line spacing purposes we look at what line the previous entity's
limit position was. By tamping it to be at the EOL instead of at the 0 position
of the next line, we get appropriate paragraph-style-spacing determination.

PiperOrigin-RevId: 577330728