Block-jit in eval_proc_main
Add a new block_jit backend to eval_proc_main. This was done by implementing a new continuation API for both the interpreter and the JIT to enable code-reuse.

This new block_jit backend is >100x the speed of the interpreter for large test cases.

The jit can run idct_chen_13000ps_model_sky130_benchmark_ir 62000 times/second on random inputs and the interpreter can run it 909 times/second. This is a 55x improvement.

PiperOrigin-RevId: 578580506