[DSLX:fmt] Make flat-or-break group associate operator w/RHS of the binop.
This lets binary operations "pack in" to lines, whereas if we associate the
operator with the left hand term, once the LHS term enters break mode, the RHS
will be broken into a new line as well as it's also in break mode.

This comes from the fact that our same-precedence-operator trees are left leaning; e.g.

A || B || C

Was grouped as:

group(group(A || B) || C)

In the case where `A || B` doesn't fit, neither the inner nor outer group could
be emitted in flat mode, so the outer group was also in break mode, and so
you'd get `|| C` on its own line even when it could pack in to the `|| B` line.

Now we group it as:

group(A group( || group(B group(|| C))))

Which creates a group where `|| B || C` can be emitted flat into a line if they
fit together.

PiperOrigin-RevId: 578962130