[block conversion]: prefix names of instantiation ports with the instantiation name.
This gives better names when there are multiple instantiations of the same module. Previously, multiple instantiations would lead to an error with name collisions, which can still happen if instance and port names alias (google/xls#1182), but this solves the common case and is more readable than our normal way of uniquifying (appending a numeric suffix).

PiperOrigin-RevId: 579338388