xls:simulation:generic: Handle back pressure from master port
This commit adds handling of the master port induced
back pressure on the streamdmachannel.

Previous integration with Renode did not require this
feature as transfers always succeed, thus no back pressure on the DMA channels.

Simulator such as gem5 can refuse to service memory access and
Require it to be retried later. This creates back pressure on the
DMA channels, and it must be handled.

Internal-tag: [#51059]

Signed-off-by: Maciej Dudek <mdudek@antmicro.com>