[DSLX:fmt] Properly handle let statement where semicolon overflows char limit.
Using the group() primitive in the pretty printing DSL is a non-trivial
decision. Forming a group means "I don't care if the outer document is in break
mode, I want to emit in flat mode if possible, even if that's the case."

But in a case like:

    let lhs = rhs;

When the RHS is in a group, but the semicolon is not, it gets shielded from the
fact the subsequent semicolon may not fit onto the line. By contrast, if we /do
not/ wrap the rhs in a group, it is exposed to whether the outer document is
currently in break mode (i.e. a document that includes the semicolon), and can
take appropriate steps to emit itself as a result.

This change removes the "group shielding" that was happening in conditionals
which led to an over-long line when it was the RHS of a let statement.

PiperOrigin-RevId: 579877945