Fix handling of `decode` operation in the JIT
Previously tried to compare the index to the bitwidth of the output, with both represented in the type of the `decode`'s result... however, since `decode`'s output width is an arbitrary parameter, it's possible for this to be too narrow to fit the index.

Also always relied on `zext` to make the shift amount fit the result type, but this is invalid LLVM IR; we instead choose between `zext` and `trunc` as appropriate.

PiperOrigin-RevId: 579917828