Add option to report per-state-element slack needed to make infeasible schedules feasible.
When our scheduler finds that the requested design is infeasible to schedule, it can reformulate the problem with slack variables that, together with a carefully crafted objective function, produce solutions that can be interpreted to provide actionable guidance as part of an error message. Currently, this guidance consists of values for the number of pipeline stages or worst-case throughput. This change adds a new flag that can produce per-node guidance, for example that `state0` getting a slack of 1 will make the schedule feasible. This is especially useful in large designs where it can be hard to understand why worst-case throughput is worse than you thought it should be.

PiperOrigin-RevId: 580638976