[DSLX:fmt] Fix the case where top-level members are multi-line and abutted.
Before we were comparing the "start" position of two adjacent top-level
members, but really you want to compare the limit position of the above one to
the start position of the below one.

This also makes some tweaks to ensure the end-delimiter's limit position is the
limit position for the Array and ConstantDef constructs.

PiperOrigin-RevId: 580653109