Allow to access tuple parameter values in FFI templates.
This allows to reference individual parts of tuple parameters
to functions in the FFI template.

Often we use tuples are a neat abstraction as parameters to
DSLX functions, but their contents might need to be accessed
separately in the FFI template to pass to verilog ports.
This change allows to reference these naturally with e.g. {a.0}.

PiperOrigin-RevId: 583410693