Add basic correctness check to pass_base for return value
The return value of a pass's Run function is supposed to indicate if anything changed. If this is incorrect it can cause significant issues. To mitigate this somewhat perform a quick correctness check by making sure that the number of nodes hasn't changed if the pass says it didn't do anything.

PiperOrigin-RevId: 583436470