[DSLX:fmt] Add "reduce text width" command to pretty print DSL.
The way the pretty printing DSL works is it greedily emits things as flat until
it can't, at which point it switches in to break mode.

If we emit something flat until column 100, then we need to insert a semicolon,
we'll need to put it on a newline when we're in break mode.

Having a semicolon on its own line is odd, so we reduce the text width
available to 99 chars in the RHS of a let expression, so that we know we will
have room to emit the semicolon inline. This dovetails with the greediness of
flat requirements.

PiperOrigin-RevId: 584349774