[DSLX:stdlib] Small tweak for readability.
Instead of packing a complex expression into the type annotation directly,
compute it in a const binding immediately above.

This was an interesting challenge for the auto-formatter, but there's a good
argument to be made that it should be written slightly differently to be more
readable -- doing lots of works in type annotations is non-ideal.

PiperOrigin-RevId: 584359282