xls/dslx/ir_conversion: Add test with implicit parametric function call by multiple sources
This commit adds a test that checks DSLX to IR conversion
when a parametric function is called by
both non parametric function and process.

The test uses IR conversion GetOrder function,
as it’s a case when IR "top" value is undefined.
This occurs when jitting or
running ir_conversion_main without specifying "top" argument.

In this test `main.next` should be a top level function,
with `call_parametric_from_no_parametric` and `parametric_fn` as callees.

Signed-off-by: Maciej Dudek <mdudek@antmicro.com>