Change channel reference in send/recive nodes from numeric id to channel name. For example:
```
 s: token = send(tok, x, channel_id=42)
```

becomes

```
 s: token = send(tok, x, channel=my_channel)
```

The proc-scoped channels refactor will have two channel constructs: a channel reference and the channel object itself.  The channel itself will have a unique numeric identifier, but the channel reference will simply have a name and be bound to channel at "elaboration" time. Send/receive nodes will refer to the channel reference which is name-only.

As a marginal benefit, the IR is more readable.

PiperOrigin-RevId: 585728854