[DSLX] Add build rule that can strip comments from .x files in a filegroup.
This helps streamline the test corpus examples that we present to the fuzzing
infrastructure -- right now many of the examples we get unnecessarily have
comment blocks at the top, this should avoid the need to eliminate those in
minimization.

Also, since we can get scanning errors, we add a flag to the
`strip_comments_main` binary that tells it to spit out the original contents in
the cases where a scan error is encountered.

PiperOrigin-RevId: 588258286