[DSLX:type_system] Avoid internal (type missing) error when parametric fn is not invoked.
The typechecking process is written such that dependencies in an operation have
their types resolved. i.e. it's deductive, you have types for the operands, you
have a rule, and then the resulting type for the operation is deduced.

However, the order in which we do inference for functions, e.g. parametric
functions, is driven by invocations, which concretize the types for the
parametric function for a given call site. When a parametric function is not
invoked, it doesn't get a type deduced/resolved in the dependency order, and we
ended up with a "type missing error".

Now, before we flag an internal type missing error, we see if the name refers
to a parametric function that's not being invoked -- in that case, we flag a
more specific (non-internal) type inference error, which should help account
for the fact it doesn't have a type resolved by the normal means. It's
definitionally a bit of an exceptional case.

Alternatively, we could do a pre-pass to find any parametric functions that are
not used in an invocation, but the system is fairly set up to add another
dynamic check vs adding a look-before-you-leap pass.

PiperOrigin-RevId: 588834999