[DSLX:type_system] Error when still-parametric type returned from non-parametric function.
Previously we supported type signatures that had types in it that were not
fully concretized (i.e. they still had unbound parametrics). This is one step
towards disallowing that.

Note we still need to disallow it on the parameters, that will be the followup,
and then we should be able to assert the invariant for all concrete functions
for their params & return type.

PiperOrigin-RevId: 589852826