Unify jit argument handling
Previously each jit implementation implemented the creation of and alignment of the jit-argument buffers individually. This could (and did) lead to issues with alignment appearing and being difficult to solve across all jits.

This change unifies all the buffer allocation behind a single JitArgumentSet API which ensures all pointers are aligned consistently.

PiperOrigin-RevId: 590976580