Make sure that ir_minimizer can inline functions which invoke other functions.
We had a ret-check that functions which are inlined do not have any invoke instructions in their code. This is a useful correctness check in general to ensure we are inlining everything in one pass but for ir_minimizer its undesirable. This changes the API the minimizer calls to disable that check.

PiperOrigin-RevId: 591031560