Use ternary query when replacing known values with literals
This is substantially faster when querying the ternary query engine, and currently cheap for intervals as well (since it's maintained automatically); an internal benchmark suite shows >10% reduced runtime for `opt_main` on average, with some large benchmarks closer to a 30% reduction.

Also optimize the case when query engines are unable to respond to a ternary query, and implement the remaining case (non-Bits nodes) for the union query engine.

PiperOrigin-RevId: 594315253