Allow opt to take explicit passes to run in a specific order
This adds an `opt --passes` flag which takes a text pass pipeline to run on the input. The grammar for this pipeline is:

```
<PIPELINE> ::= <NAME> | <NAME> <PIPELINE> | FIXEDPOINT
<FIXEDPOINT> := '[' <PIPELINE> ']'
```

Symbols break on whitespace unconditionally. For opt the names are the short_names with configuration contained in '()' where appropriate. Other tools can interpret the <NAME> token however they wish however.

FIXEDPOINT elements are run to fixedpoint, care should be taken that the pipeline will finish in an acceptable amount of time.

An example pipeline could be

```
opt --passes 'inlining dce arith_simp [dce table_switch narrow(Bdd) dce select_simp]'
```

This is similar to how the llvm `opt` and similar tools can take a list of passes to run on the code.

This feature is meant for debugging issues with specific passes or groups of passes only.

PiperOrigin-RevId: 596755396